Move the concept of iterations out of the sample computation logic
This was a failed attempt at playing multiple single link failure
to avoid picking a sensitive link by chance. The bottleneck of the
loading speed of the browser, however, was the computation of routes,
not the computation of samples. Therefore, playing multiple events
at once ended up freezing the browser for large topologies.

Nevertheless, we keep the changes because it allows for more flexibility.